% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMShiny.R
\name{CMShiny}
\alias{CMShiny}
\title{Launch 'CMShiny' Interface}
\usage{
CMShiny()
}
\value{
Nothing
}
\description{
CMShiny() loads interactive user interface built using R 'shiny'.
}
\details{
Start CMShiny

The interactive user interface is to provide an easy way for confusion matrix analysis and downloading relevant plot and dataset.
}
\examples{
if(interactive()){
library(rmarkdown)
CMShiny()
}
}
\keyword{CMShiny}
